Removed DecayMegacomplexBase
