+++ glotaran/builtin/megacomplexes/decay/decay_parallel_megacomplex.py
@@ -4,11 +4,15 @@ from __future__ import annotations
 from typing import List
 
 import numpy as np
+import xarray as xr
 
-from glotaran.builtin.megacomplexes.decay.decay_megacomplex_base import DecayMegacomplexBase
 from glotaran.builtin.megacomplexes.decay.irf import Irf
 from glotaran.builtin.megacomplexes.decay.k_matrix import KMatrix
+from glotaran.builtin.megacomplexes.decay.util import calculate_matrix
+from glotaran.builtin.megacomplexes.decay.util import finalize_data
+from glotaran.builtin.megacomplexes.decay.util import index_dependent
 from glotaran.model import DatasetModel
+from glotaran.model import Megacomplex
 from glotaran.model import megacomplex
 from glotaran.parameter import Parameter
 
@@ -24,7 +28,7 @@ from glotaran.parameter import Parameter
     },
     register_as="decay-parallel",
 )
-class DecayParallelMegacomplex(DecayMegacomplexBase):
+class DecayParallelMegacomplex(Megacomplex):
     def get_compartments(self, dataset_model: DatasetModel) -> list[str]:
         return self.compartments
 
@@ -38,3 +42,35 @@ class DecayParallelMegacomplex(DecayMegacomplexBase):
             (self.compartments[i], self.compartments[i]): self.rates[i] for i in range(size)
         }
         return k_matrix
+
+    def index_dependent(self, dataset_model: DatasetModel) -> bool:
+        return index_dependent(dataset_model)
+
+    def calculate_matrix(
+        self,
+        dataset_model: DatasetModel,
+        indices: dict[str, int],
+        **kwargs,
+    ):
+        return calculate_matrix(self, dataset_model, indices, **kwargs)
+
+    def finalize_data(
+        self,
+        dataset_model: DatasetModel,
+        dataset: xr.Dataset,
+        is_full_model: bool = False,
+        as_global: bool = False,
+    ):
+        from glotaran.builtin.megacomplexes.decay.decay_megacomplex import DecayMegacomplex
+        from glotaran.builtin.megacomplexes.decay.decay_sequential_megacomplex import (
+            DecaySequentialMegacomplex,
+        )
+
+        decay_megacomplexes = [
+            m
+            for m in dataset_model.megacomplex
+            if isinstance(
+                m, (DecayMegacomplex, DecayParallelMegacomplex, DecaySequentialMegacomplex)
+            )
+        ]
+        finalize_data(decay_megacomplexes, dataset_model, dataset, is_full_model, as_global)
